requirements.txt
# Core dependencies
numpy>=1.23.5,<2.0.0
pandas>=1.5.3,<2.0.0
PyYAML>=6.0,<7.0.0
python-dateutil>=2.8.2,<3.0.0
pytz>=2020.1,<2026.0
jsonschema>=4.17.3,<5.0.0

# Excel reporting
XlsxWriter>=3.0.0,<4.0.0
openpyxl>=3.0.0,<4.0.0

# Technical analysis
ta>=0.10.0,<0.12.0

# Visualization
matplotlib>=3.5.0,<4.0.0

# Reporting
tabulate>=0.9.0,<1.0.0

# Performance optimization
numba>=0.56.0,<0.61.0

# Logging and CLI
rich>=14.0.0,<15.0.0
argparse>=1.4.0,<2.0.0

# Testing
pytest>=7.3.1,<9.0.0
pytest-cov>=4.1.0,<7.0.0
